Fix printf 'implicit conversion from float to double' errors.
These warnings showed up when enabling the '-Wdouble-promotion' GCC
compiler flag.  This batch of fixes consists of fixing arguments to
printf statements.  The printf e/f/g conversion specifiers all take a
double as their argument.  Explicitly convert floats to doubles to
eliminate these errors.

+++ mythplugins/mythmusic/mythmusic/musiccommon.cpp
@@ -1316,13 +1316,13 @@ void MusicCommon::customEvent(QEvent *event)
         if (oe->bitrate() < 2000)
         {
             info_string.sprintf(QString("%d "+tr("kbps")+ "   %.1f "+ tr("kHz")+ "   %s "+ tr("ch")).toUtf8().data(),
-                                oe->bitrate(), float(oe->frequency()) / 1000.0,
+                                oe->bitrate(), static_cast<double>(oe->frequency()) / 1000.0,
                                 oe->channels() > 1 ? "2" : "1");
         }
         else
         {
             info_string.sprintf(QString("%.1f "+ tr("kHz")+ "   %s "+ tr("ch")).toUtf8().data(),
-                                float(oe->frequency()) / 1000.0,
+                                static_cast<double>(oe->frequency()) / 1000.0,
                                 oe->channels() > 1 ? "2" : "1");
         }
